Separate enhanced sampling objects into their own files. Delete analysis file.
